extract config models from report
